+++ src/libraries/Common/src/System/IO/Win32Marshal.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using System.Runtime.InteropServices;
-using Microsoft.Win32.SafeHandles;
 
 namespace System.IO
 {
